Move htmlspecialchars__recursive and htmltrim__recursive to the Utils class + move some other functions to Subs.php
